Moved DiscoverdInfo to Configuration object.
Implemented injection for IActivityMonitor and IYodiiEngine constructor parameters.
An ActivityMonitor is now created for each Plugin.
IYodiiEngineExternal is the external view of the Engine.
IYodiiEngine is not INPC at its level: its Configuration and LiveInfo are.
This interface is used to inject the callerKey when Start/Stop are called from a plugin (the callerKey is the plugin full name).