Fixed errors when running tests on external NUnit (...)
Caused by problematic struct when calling Win32 APIs.
WPF-related tests are now explicit.
Note that, currently, the UI thread (and WPF Application) is never shut down, as it cannot be started again in the same AppDomain.
It would be nice to be able to hook to when NUnit closes to dispose of the Application and Thread, but I don't know of any way to do just that.